+++ src/authzee/grant_effect.py
@@ -3,6 +3,5 @@ from enum import Enum
 
 
 class GrantEffect(Enum):
-
     ALLOW = "ALLOW"
     DENY = "DENY"