+++ src/authzee/page_refs_page.py
@@ -1,0 +1,9 @@
+
+from typing import List, Union
+
+from pydantic import BaseModel
+
+
+class PageRefsPage(BaseModel):
+    page_refs: List[str]
+    next_page_ref: Union[str, None]