+++ src/authzee/raw_grants_page.py
@@ -5,7 +5,6 @@ from pydantic import BaseModel
 
 
 class RawGrantsPage(BaseModel):
-
     raw_grants: Any
     next_page_ref: Union[str, None]