+++ CHANGELOG.md
@@ -43,8 +43,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### [0.1.0a3] - 2024-02-DD
 
 - Remove sync methods and only accepts async
-- Locality checks streamlined.
-- add needed boilerplate for parallel paging
+- `AuthzeeSync` class as a sync wrapper for Authzee app
+- Locality checks streamlined. Only specify current backends locality
+- Parallel pagination skeleton
 - general renaming
 - rework ResourceAuthzs. They are now passed as instances to the Authzee app. Logic is streamlined